🆙update: StatusIndicator - 修改 vue (檢整 showCode)   1. 更改資料夾名稱 - 修改 storybook (檢整 showCode)   1. 改寫 showCode
